+++ apps/cms/config/plugins.ts
@@ -1,0 +1,11 @@
+export default ({
+  env,
+}: {
+  env: (key: string, fallback?: unknown) => unknown;
+}) => ({
+  "users-permissions": {
+    config: {
+      jwtSecret: env("JWT_SECRET"),
+    },
+  },
+});